# speed up post-processing