Also max out detections during benchmarking to stress test that part

Before speedup
```
┌──────────────────┬─────────┬───────────────┬──────────┬──────────────┬────────────────┬─────────────┐
│            Model │ loaded? │ load time (s) │ #results │ run time (s) │ run time (fps) │ allocations │
├──────────────────┼─────────┼───────────────┼──────────┼──────────────┼────────────────┼─────────────┤
│ v2_tiny_416_COCO │    true │         2.701 │      845 │       0.0501 │           19.9 │   3.331 MiB │
│ v3_tiny_416_COCO │    true │         0.375 │     2535 │       0.0674 │           14.8 │  16.641 MiB │
│      v3_320_COCO │    true │         1.424 │     6300 │       0.3057 │            3.3 │  64.624 MiB │
│      v3_416_COCO │    true │         0.896 │    10647 │       0.5563 │            1.8 │ 177.833 MiB │
│      v3_608_COCO │    true │         1.445 │    22743 │       2.1043 │            0.5 │   1.222 GiB │
└──────────────────┴─────────┴───────────────┴──────────┴──────────────┴────────────────┴─────────────┘
```

After speedup
```
┌──────────────────┬─────────┬───────────────┬──────────┬──────────────┬────────────────┬─────────────┐
│            Model │ loaded? │ load time (s) │ #results │ run time (s) │ run time (fps) │ allocations │
├──────────────────┼─────────┼───────────────┼──────────┼──────────────┼────────────────┼─────────────┤
│ v2_tiny_416_COCO │    true │         2.734 │      845 │       0.0442 │           22.6 │ 697.438 KiB │
│ v3_tiny_416_COCO │    true │         0.385 │     2535 │       0.0476 │           21.0 │   1.903 MiB │
│      v3_320_COCO │    true │         1.501 │     6300 │       0.2582 │            3.9 │   4.844 MiB │
│      v3_416_COCO │    true │         0.788 │    10647 │       0.4195 │            2.4 │   7.849 MiB │
│      v3_608_COCO │    true │         1.345 │    22743 │       1.0851 │            0.9 │  16.347 MiB │
└──────────────────┴─────────┴───────────────┴──────────┴──────────────┴────────────────┴─────────────┘
```

```
julia> @time size(yolomod(batch, detectThresh=0.0, overlapThresh=1.0, show_timing=true))
───────────────────────────────────────────────────────────────────────────────────
                                          Time                    Allocations
                                 ───────────────────────   ────────────────────────
        Tot / % measured:             857ms / 100.0%           8.65MiB / 100.0%

Section                  ncalls     time    %tot     avg     alloc    %tot      avg
───────────────────────────────────────────────────────────────────────────────────
yolo                          1    857ms  100.0%   857ms   8.65MiB  100.0%  8.65MiB
  forward pass                1    712ms   83.1%   712ms    410KiB    4.6%   410KiB
    layer 1                   1   81.4ms    9.5%  81.4ms   5.56KiB    0.1%  5.56KiB
    layer 2                   1   73.4ms    8.6%  73.4ms   9.22KiB    0.1%  9.22KiB
    layer 3                   1   30.1ms    3.5%  30.1ms   6.53KiB    0.1%  6.53KiB
    layer 4                   1   46.0ms    5.4%  46.0ms   11.0KiB    0.1%  11.0KiB
    layer 5                   1   18.5ms    2.2%  18.5ms   8.09KiB    0.1%  8.09KiB
    layer 6                   1   18.9ms    2.2%  18.9ms   8.09KiB    0.1%  8.09KiB
    layer 7                   1   24.2ms    2.8%  24.2ms   8.09KiB    0.1%  8.09KiB
    layer 8                   1   18.4ms    2.1%  18.4ms   8.09KiB    0.1%  8.09KiB
    layer 9                   1   17.9ms    2.1%  17.9ms   8.09KiB    0.1%  8.09KiB
    layer 10                  1   18.0ms    2.1%  18.0ms   8.09KiB    0.1%  8.09KiB
    layer 11                  1   18.1ms    2.1%  18.1ms   8.09KiB    0.1%  8.09KiB
    layer 12                  1   18.3ms    2.1%  18.3ms   8.09KiB    0.1%  8.09KiB
    layer 13                  1   11.6ms    1.4%  11.6ms   6.38KiB    0.1%  6.38KiB
    layer 14                  1   13.7ms    1.6%  13.7ms   11.1KiB    0.1%  11.1KiB
    layer 15                  1   13.4ms    1.6%  13.4ms   11.1KiB    0.1%  11.1KiB
    layer 16                  1   13.5ms    1.6%  13.5ms   11.1KiB    0.1%  11.1KiB
    layer 17                  1   13.3ms    1.6%  13.3ms   11.1KiB    0.1%  11.1KiB
    layer 18                  1   13.7ms    1.6%  13.7ms   11.1KiB    0.1%  11.1KiB
    layer 19                  1   13.5ms    1.6%  13.5ms   11.1KiB    0.1%  11.1KiB
    layer 20                  1   13.3ms    1.5%  13.3ms   11.1KiB    0.1%  11.1KiB
    layer 21                  1   13.2ms    1.5%  13.2ms   11.1KiB    0.1%  11.1KiB
    layer 22                  1   12.2ms    1.4%  12.2ms   10.4KiB    0.1%  10.4KiB
    layer 23                  1   14.0ms    1.6%  14.0ms   17.0KiB    0.2%  17.0KiB
    layer 24                  1   14.2ms    1.7%  14.2ms   17.0KiB    0.2%  17.0KiB
    layer 25                  1   14.0ms    1.6%  14.0ms   17.0KiB    0.2%  17.0KiB
    layer 26                  1   41.1ms    4.8%  41.1ms   57.2KiB    0.6%  57.2KiB
    layer 27                  1   13.2ms    1.5%  13.2ms   11.7KiB    0.1%  11.7KiB
    layer 28                  1   27.8ms    3.2%  27.8ms   32.6KiB    0.4%  32.6KiB
    layer 29                  1   13.4ms    1.6%  13.4ms   7.72KiB    0.1%  7.72KiB
    layer 30                  1   59.6ms    6.9%  59.6ms   30.3KiB    0.3%  30.3KiB
  post-processing             1    145ms   16.9%   145ms   8.25MiB   95.4%  8.25MiB
    processing outputs        1   22.7ms    2.6%  22.7ms   73.1KiB    0.8%  73.1KiB
    filter detections         1   3.14ms    0.4%  3.14ms   3.23KiB    0.0%  3.23KiB
    nms                       1    119ms   13.9%   119ms   8.17MiB   94.5%  8.17MiB
───────────────────────────────────────────────────────────────────────────────────
  0.934027 seconds (238.41 k allocations: 16.503 MiB)
(89, 22743)
```