@IanButterworth 2025-03-27
Tests pass on a CUDA machine